// pratik/pratik/f/src/interfaces/cari-interface.ts
export interface Cari {
  CariID: number;
  CariKod: string;
  CariIsim: string;
  Il: string;
  Ilce: string;
  VergiDairesi: string;
  VergiNumarasi: string;
  TCKimlikNo: string;
  CariRol: string;
  KayitTarihi: string;
  GrupKodu1Isim: string;
  GrupKodu2Isim: string;
  GrupKodu3Isim: string;
} 